refactor: 오늘의시간 페이지 컴포넌트 기반 리팩토링
